+++ consent/consent-page/external_consent_storage.go
@@ -13,30 +13,7 @@ import (
 	"os"
 
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
-
-	"github.com/cloudentity/acp-client-go/clients/system/models"
 )
-
-type ConsentID string
-
-type Subject string
-
-type Data models.ScopeGrantSessionResponse
-
-type ConsentStorage interface {
-	Store(ctx context.Context, sub Subject, data Data) (ConsentID, error)
-}
-
-type DummyConsentStorage struct{}
-
-var _ ConsentStorage = &DummyConsentStorage{}
-
-func (d DummyConsentStorage) Store(ctx context.Context, sub Subject, data Data) (ConsentID, error) {
-	logrus.Infof("Store consent for sub: %s with data: %+v", sub, data)
-
-	return "external-consent-id", nil
-}
 
 type ExternalConsentStorage struct {
 	URL *url.URL
